server/pyproject.toml
[project]
name = "autholau-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.3",
    "flask-cors==4.0.1",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.1",
]
